+++ kedro_mlflow/io/__init__.py
@@ -1,3 +1,2 @@
 from .mlflow_dataset import MlflowArtifactDataSet
 from .mlflow_metrics_dataset import MlflowMetricsDataSet
-from .mlflow_model_dataset import MlflowModelDataSet